requirements.txt
-e ./OFA/fairseq/
torch==1.12.1
transformers==4.2.1
sentencepiece
h5py
wandb
tqdm
numpy
pandas
matplotlib
pyyaml
sacrebleu
torchvision==0.13.1
wget
 
timm
ftfy==6.0.3
tensorboardX==2.4.1
pycocotools==2.0.4
pycocoevalcap==1.2
pytorch_lightning
einops
datasets
rouge_score
soundfile
editdistance
librosa
python-Levenshtein
zhconv
pypinyin==0.47.1